Splits flock.midi.controller out into its own file.
