+++ demos/playground/index.html
@@ -66,6 +66,7 @@
         <script src="../../src/web/buffer-writer.js"></script>
         <script src="../../src/web/input-device-manager.js"></script>
         <script src="../../src/web/midi.js"></script>
+        <script src="../../src/midi/controller.js"></script>
         <script src="../../src/web/native-node-manager.js"></script>
         <script src="../../src/web/output-manager.js"></script>
         <script src="../../src/parser.js"></script>
