+++ demos/playground/live/index.html
@@ -70,6 +70,7 @@
         <script src="../../../src/web/buffer-writer.js"></script>
         <script src="../../../src/web/input-device-manager.js"></script>
         <script src="../../../src/web/midi.js"></script>
+        <script src="../../../src/midi/controller.js"></script>
         <script src="../../../src/web/native-node-manager.js"></script>
         <script src="../../../src/web/output-manager.js"></script>
         <script src="../../../src/parser.js"></script>
